# tmp-for-review/ericsson/teiv-seedcode/teiv/src/main/resources/v1/topology-exposure-inventory-openapi.yaml
...
openapi: 3.0.2
info:
  x-api-id: 52812f69-83ac-4dfa-b83a-1a1bdf2d49b8
  x-audience: external-public
  description: |
    Topology and Inventory data is the information that represents entities in a telecommunications network and the relationships between them that provide insight into a particular aspect of the network of importance to specific use cases. Topology and Inventory data can be derived from inventory, configuration, or other data.

    Topology and Inventory supports several topology domains. A domain is a grouping of network topology entities which handles topology data.

    Entities are enabling the modelling and storage of complex network infrastructure and relationships.

    Relationships are a uni-directional connection between two entities, one of which is the originating side (A-side) and the other is the terminating side (B-side). The order of the sides matters since it defines the relationship itself which must be unique

    Topology and Inventory API provides the capabilities to fetch topology data. Using the filtering options, it is possible to define more specific query requests.

    ## Querying simple entities
...

...

    ## Querying connected entities
    It is possible to get information about directly connected objects as well. If entityType is present in the *targetFilter* parameter, the query provides information about that entityType itself.

...

  version: 0.11.0
  title: Topology and Inventory API
  license:
...
servers:
...
    variables:
      host:
...
paths:
...
  /schemas:
    post:
      description: Create a new schema.
      tags:
        - Schemas
      summary: Create a new schema.
      operationId: createSchema
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/contentTypeInHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultipartFile'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      description: Get a list of all schemas.
      tags:
        - Schemas
      summary: Get a list of all schemas.
      operationId: getSchemas
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/domainOptionalInQuery'
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaList'
              examples:
                schemas:
                  $ref: '#/components/examples/SchemasResponseExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /schemas/{schemaName}/content:
    get:
      description: Get the model schema by name.
      tags:
        - Schemas
      summary: Get the model schema.
      operationId: getSchemaByName
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/schemaNameInPath'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              examples:
                schema:
                  $ref: '#/components/examples/SchemaResponseExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /schemas/{schemaName}:
    delete:
      description: Delete a schema.
      tags:
        - Schemas
      summary: Delete a schema.
      operationId: deleteSchema
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/schemaNameInPath'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    ErrorMessage:
      type: object
      title: Error
      properties:
        status:
          type: string
        message:
          type: string
        details:
          type: string
    Href:
      type: object
      title: Href
      properties:
        href:
          type: string
          format: uri-template
    MultipartFile:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          description: multipartFile
          format: binary
...
    Schema:
      type: object
      title: Schema
      properties:
        name:
          type: string
        domain:
          type: array
          items:
            type: string
        revision:
          type: string
        content:
          $ref: '#/components/schemas/Href'
    SchemaList:
      type: object
      title: Schemas
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Schema'
        self:
          $ref: '#/components/schemas/Href'
        first:
          $ref: '#/components/schemas/Href'
        prev:
          $ref: '#/components/schemas/Href'
        next:
          $ref: '#/components/schemas/Href'
        last:
          $ref: '#/components/schemas/Href'
        totalCount:
          type: integer
  responses:
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          example:
            status: '404'
            title: Resource Not Found
            details: The requested resource is not found
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          example:
            status: '401'
            title: Unauthorized request
            details: This request is unauthorized
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          example:
            status: '403'
            title: Request Forbidden
            details: This request is forbidden
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          example:
            status: '400'
            title: Bad Request
            details: The provided request is not valid
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          example:
            status: 409'
            title: Conflicting request
            details: The request cannot be processed as the resource is in use.
    Created:
      description: Created without response body
    InternalServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
          example:
            status: '500'
            title: Internal Server Error
            details: Internal Server Error occurred
    NoContent:
      description: No Content
      content: {}
  parameters:
    acceptInHeader:
      name: Accept
      in: header
      required: true
      schema:
        type: string
        example: application/json
        default: application/json
    contentTypeInHeader:
      name: Content-Type
      in: header
      required: true
      schema:
        type: string
        example: application/json
        default: application/json
    offsetParam:
      name: offset
      in: query
      description: Pagination offset.
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    limitParam:
      name: limit
      in: query
      description: Result limiter.
      required: false
      schema:
        type: integer
        default: 500
        minimum: 1
        maximum: 500
...
    schemaNameInPath:
      name: schemaName
      in: path
      required: true
      schema:
        type: string
        default: "o-ran-smo-teiv-ran"
...
    domainOptionalInQuery:
      name: domain
      in: query
      required: false
      schema:
        type: string
      examples:
        domain:
          value: ran
...
  examples:
...
    SchemasResponseExample:
      value:
        items:
          - name: o-ran-smo-teiv-ran
            domain:
              - RAN_LOGICAL
            revision: '2013-07-15'
            content:
              href: '/schemas/o-ran-smo-teiv-ran/content'
          - name: o-ran-smo-teiv-cloud
            domain:
              - RAN_CLOUD
            revision: '2021-05-19'
            content:
              href: '/schemas/o-ran-smo-teiv-cloud/content'
          - name: o-ran-smo-teiv-oam
            domain:
              - RAN_OAM'
            revision: '2021-05-19'
            content:
              href: '/schemas/o-ran-smo-teiv-oam/content'
          - name: o-ran-smo-teiv-common-yang-types
            domain:
              []
            revision: '2021-07-04'
            content:
              href: '/schemas/o-ran-smo-teiv-common-yang-types/content'
          - name: o-ran-smo-teiv-common-yang-extensions
            domain:
              []
            revision: '2021-07-04'
            content:
              href: '/schemas/o-ran-smo-teiv-common-yang-extensions/content'
        self:
          href: '/schemas?offset=0&limit=500'
        first:
          href: '/schemas?offset=0&limit=500'
        prev:
          href: '/schemas?offset=0&limit=500'
        next:
          href: '/schemas?offset=500&limit=500'
        last:
          href: '/schemas?offset=678&limit=500'
        totalCount: 14
    SchemaResponseExample:
      value: "module stores {\
  yang-version 1.1;\
  namespace \"ietf-inet-types\";\
  prefix module-references;\
  revision \"2020-09-15\" {\
    description\
    \"Sample Model\";\
    }  \
  typedef Mcc {\
    type string;\
    }\
  typedef Mnc {\
    type string;\
    }\
  grouping PLMNId {\
    leaf mcc {\
      type Mcc;\
        }\
    leaf mnc {\
      type Mnc;\
        }\
    }      \
  container RAN_LOGICAL {\
    leaf namespace {\
      type string;\
        }\
    container GNBDUFunction {\
      leaf eiid {\
        type string;\
            }  \
      leaf gNBDUId {\
        type uint16;\
            }\
      leaf gNBID {\
        type uint16;\
            }\
      leaf gNBIdLength {\
        type uint16;\
            }\
      container dUpLMNId{\
        description \"ToDo\";\
        uses PLMNId;\
      }        \
    }\
    container GNBCUUPFunction {\
      leaf eiid {\
        type string;\
      } \
      leaf gNBID {\
        type uint16;\
            }\
      leaf gNBIdLength {\
        type uint16;\
            }\
        }  \
    container GNBCUCPFunction {            \
      leaf eiid {\
        type string;\
            } \
      leaf gNBID {\
        type uint16;\
            }\
      leaf gNBIdLength {\
        type uint16;\
            }\
      leaf gNBCUName {\
        type string;\
            }\
      container pLMNId{\
        description \"ToDo\";\
        uses PLMNId;\
      }      \
        }\
    container NRCellCU {            \
      leaf eiid {\
        type string;\
            } \
      leaf cellLocalId {\
        type uint16;\
            }\
      leaf nCI {\
        type uint16;\
            }\
      leaf nRTAC {\
        type uint16;\
            }\
      container plmnId{\
        description \"ToDo\";\
        uses PLMNId;\
      }        \
        }\
    container NRCellDU {            \
      leaf eiid {\
        type string;\
            } \
      leaf cellLocalId {\
        type uint16;\
            }\
      leaf nCI {\
        type uint16;\
            }\
      leaf nRPCI {\
        type uint16;\
            }\
      leaf nRTAC {\
        type uint16;\
            }\
        }\
    }\
}"